combat, doors open and close
i made the combat mechanics work on the new script of player movement, the room collider now can activate and deactivate rooms based on the objects of the layer Enemy present on it. Tried a automatical camera code but didnt finish it. the rooms already visited work fine but the adjacent are not working.